warning if not building disk as root

+++ build_disk.sh
@@ -1,4 +1,9 @@
 #!/bin/sh
+if [ "$EUID" -ne 0 ]
+then
+    echo "Please run this script as root."
+    exit
+fi
 mkfs.ext4 -F disk
 mount disk mnt
 cd mnt
